requirements.txt
fastapi~=0.103.0
uvicorn~=0.20.0
sqlalchemy~=2.0.30

paho-mqtt~=2.1.0
pydantic~=2.5.3